Provided a way to set the serial port baud rate.
By letting PlatformSerialP provide the Msp430UsciConfigure interface and implement the getConfiguration method, we can
now pass a different msp430_usci_config_t variable instead of using the default. This is necessary for the EXP430
because the default baud rate table (msp430usci.h) assumes the UART clock input(SMCLK) to be 1MiHz instead of the
correct value of 8MHz.