Implemented Clear Channel assessment.
The CC1101 can perform CCA when transitioning from RX to TX state. This is called TX-if-CCA. With this enabled,
RadioCCA.request() returns immediately with SUCCESS and the actual CCA is postponed until RadioSend.send() is called.
Furthermore, IOCFG0 is configured to 0x06 in which GDO0 will assert when the sync word has been sent/received. This way,
we don't have to poll the chip after filling the TXFIFO to see if it has finished transmitting.