Fixed a bug where the current time as it was read from the timer register was corrupt.
The solution was to assume the timer clock and the CPU clock were asynchronous, in which case a the timer register is
sampled multiple times and a majority vote is taken. This bug affected the radio stack by firing an alarm when it wasn't
supposed to. This in turn exposed a different bug where sending data over the radio was not functioning after resetRx
was called.